resolved import path issues created by moving page components
